Split oversized test modules
Move HTML and inline postprocess tests into test-only child modules so the
production source files stay below the repository line limit while retaining
access to private helpers.

Lower the HTML table conversion event from `warn!` to `debug!` because normal
conversions should not be reported as warnings.
