+++ src/wrap/inline/postprocess_tests.rs
@@ -1,0 +1,238 @@
+//! Unit tests for inline post-wrap normalisation.
+//!
+//! These tests compile as a child module of `postprocess`, so they can cover
+//! private helpers while keeping test-only cases out of the production module.
+
+use super::{
+    super::fragment::{FragmentKind, InlineFragment},
+    *,
+};
+
+fn fragment(text: &str) -> InlineFragment { InlineFragment::new(text.into()) }
+
+#[test]
+fn inline_fragment_whitespace_space() {
+    let fragment = InlineFragment::new(" ".into());
+    assert_eq!(fragment.kind, FragmentKind::Whitespace);
+    assert!(fragment.is_whitespace());
+    assert!(!fragment.is_atomic());
+    assert_eq!(fragment.width, 1);
+}
+
+#[test]
+fn inline_fragment_whitespace_tab() {
+    let fragment = InlineFragment::new("\t".into());
+    assert_eq!(fragment.kind, FragmentKind::Whitespace);
+}
+
+#[test]
+fn inline_fragment_inline_code() {
+    let fragment = InlineFragment::new("`foo`".into());
+    assert_eq!(fragment.kind, FragmentKind::InlineCode);
+    assert!(fragment.is_atomic());
+    assert!(!fragment.is_whitespace());
+    assert!(!fragment.is_plain());
+}
+
+#[test]
+fn inline_fragment_link() {
+    let fragment = InlineFragment::new("[text](url)".into());
+    assert_eq!(fragment.kind, FragmentKind::Link);
+    assert!(fragment.is_atomic());
+}
+
+#[test]
+fn inline_fragment_plain() {
+    let fragment = InlineFragment::new("word".into());
+    assert_eq!(fragment.kind, FragmentKind::Plain);
+    assert!(fragment.is_plain());
+    assert!(!fragment.is_atomic());
+}
+
+#[test]
+fn merge_keeps_content_lines_unchanged() {
+    let lines = vec![vec![fragment("hello")], vec![fragment("world")]];
+    assert_eq!(merge_whitespace_only_lines(&lines), lines);
+}
+
+#[test]
+fn merge_carries_whitespace_forward() {
+    let lines = vec![
+        vec![fragment("hello")],
+        vec![fragment(" ")],
+        vec![fragment("world")],
+    ];
+    assert_eq!(
+        merge_whitespace_only_lines(&lines),
+        vec![
+            vec![fragment("hello")],
+            vec![fragment(" "), fragment("world")],
+        ]
+    );
+}
+
+#[test]
+fn merge_moves_inline_code_tail_before_single_space() {
+    let lines = vec![
+        vec![fragment("plain"), fragment("`code`")],
+        vec![fragment(" ")],
+        vec![fragment("tail")],
+    ];
+    let merged = merge_whitespace_only_lines(&lines);
+
+    assert_eq!(merged[0], vec![fragment("plain")]);
+    assert_eq!(
+        merged[1],
+        vec![fragment("`code`"), fragment(" "), fragment("tail")]
+    );
+}
+
+#[test]
+fn merge_trailing_whitespace_appended_to_last_line() {
+    let lines = vec![vec![fragment("hello")], vec![fragment(" ")]];
+    assert_eq!(
+        merge_whitespace_only_lines(&lines),
+        vec![vec![fragment("hello"), fragment(" ")]]
+    );
+}
+
+#[test]
+fn merge_carries_multiple_consecutive_whitespace_lines_forward() {
+    let lines = vec![
+        vec![fragment("hello")],
+        vec![fragment(" ")],
+        vec![fragment("\t")],
+        vec![fragment("world")],
+    ];
+    assert_eq!(
+        merge_whitespace_only_lines(&lines),
+        vec![
+            vec![fragment("hello")],
+            vec![fragment(" "), fragment("\t"), fragment("world")]
+        ]
+    );
+}
+
+#[test]
+fn merge_drops_single_space_before_atomic_starting_line() {
+    let lines = vec![
+        vec![fragment("alpha"), fragment("beta")],
+        vec![fragment(" ")],
+        vec![fragment("`code`")],
+    ];
+    assert_eq!(
+        merge_whitespace_only_lines(&lines),
+        vec![
+            vec![fragment("alpha"), fragment("beta")],
+            vec![fragment("`code`")]
+        ]
+    );
+}
+
+#[test]
+fn merge_empty_input_returns_empty_output() {
+    assert!(merge_whitespace_only_lines(&[]).is_empty());
+}
+
+#[test]
+fn rebalance_moves_atomic_tail_when_fits() {
+    let mut lines = vec![
+        vec![fragment("alpha"), fragment("`code`")],
+        vec![fragment(" "), fragment("tail")],
+    ];
+    rebalance_atomic_tails(&mut lines, 80);
+    assert_eq!(lines[0], vec![fragment("alpha")]);
+    assert_eq!(
+        lines[1],
+        vec![fragment("`code`"), fragment(" "), fragment("tail")]
+    );
+}
+
+#[test]
+fn rebalance_does_not_move_when_overflow() {
+    let mut lines = vec![
+        vec![fragment("alpha"), fragment("`code`")],
+        vec![fragment(" "), fragment("plain")],
+    ];
+    let original = lines.clone();
+    let width = line_width(&lines[1]) + lines[0].last().expect("tail exists").width - 1;
+    rebalance_atomic_tails(&mut lines, width);
+    assert_eq!(lines, original);
+}
+
+#[test]
+fn rebalance_skips_when_next_does_not_start_with_space_then_plain() {
+    let mut lines = vec![
+        vec![fragment("alpha"), fragment("`code`")],
+        vec![fragment("plain")],
+    ];
+    let original = lines.clone();
+    rebalance_atomic_tails(&mut lines, 80);
+    assert_eq!(lines, original);
+}
+
+#[test]
+fn rebalance_moves_atomic_tail_at_exact_width_boundary() {
+    let mut lines = vec![
+        vec![fragment("alpha"), fragment("`tail`")],
+        vec![fragment(" "), fragment("beta")],
+    ];
+    let width = line_width(&lines[1]) + lines[0].last().expect("tail exists").width;
+    rebalance_atomic_tails(&mut lines, width);
+    assert_eq!(lines[0], vec![fragment("alpha")]);
+    assert_eq!(
+        lines[1],
+        vec![fragment("`tail`"), fragment(" "), fragment("beta")]
+    );
+}
+
+#[test]
+fn rebalance_moves_plain_tail_when_fits() {
+    let mut lines = vec![
+        vec![fragment("alpha"), fragment("tail")],
+        vec![fragment(" "), fragment("beta")],
+    ];
+    rebalance_atomic_tails(&mut lines, 20);
+    assert_eq!(lines[0], vec![fragment("alpha")]);
+    assert_eq!(
+        lines[1],
+        vec![fragment("tail"), fragment(" "), fragment("beta")]
+    );
+}
+
+#[test]
+fn rebalance_leaves_single_plain_fragment_line_unchanged() {
+    let mut lines = vec![
+        vec![fragment("tail")],
+        vec![fragment(" "), fragment("beta")],
+    ];
+    let original = lines.clone();
+    rebalance_atomic_tails(&mut lines, 20);
+    assert_eq!(lines, original);
+}
+
+#[test]
+fn rebalance_ignores_empty_input() {
+    let mut lines: Vec<Vec<InlineFragment>> = Vec::new();
+    rebalance_atomic_tails(&mut lines, 10);
+    assert!(lines.is_empty());
+}
+
+#[test]
+fn rebalance_leaves_single_line_input_unchanged() {
+    let mut lines = vec![vec![fragment("alpha"), fragment("tail")]];
+    let original = lines.clone();
+    rebalance_atomic_tails(&mut lines, 10);
+    assert_eq!(lines, original);
+}
+
+#[test]
+fn rebalance_skips_when_next_line_starts_with_space_then_atomic() {
+    let mut lines = vec![
+        vec![fragment("alpha"), fragment("tail")],
+        vec![fragment(" "), fragment("`code`")],
+    ];
+    let original = lines.clone();
+    rebalance_atomic_tails(&mut lines, 20);
+    assert_eq!(lines, original);
+}
